maj des controllers pour la verif des upload + gestion dynamiques des co en bdd

+++ src/Controller/Admin/AlbumController.php
@@ -9,8 +9,11 @@ use App\Repository\MediaRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
+#[IsGranted('ROLE_ADMIN')]
 #[Route('/admin/album')]
 class AlbumController extends AbstractController
 {
@@ -20,16 +23,18 @@ class AlbumController extends AbstractController
         private EntityManagerInterface $em
     ) {}
 
-    #[Route('/admin/album', name: 'admin_album_index')]
-    public function index()
+    #[Route('/', name: 'admin_album_index')]
+    public function index(): Response
     {
         $albums = $this->albumRepository->findAll();
 
-        return $this->render('admin/album/index.html.twig', ['albums' => $albums]);
+        return $this->render('admin/album/index.html.twig', [
+            'albums' => $albums,
+        ]);
     }
 
     #[Route('/add', name: 'admin_album_add')]
-    public function add(Request $request)
+    public function add(Request $request): Response
     {
         $album = new Album();
         $form = $this->createForm(AlbumType::class, $album);
@@ -39,34 +44,66 @@ class AlbumController extends AbstractController
             $this->em->persist($album);
             $this->em->flush();
 
+            $this->addFlash('success', 'L’album a bien été ajouté.');
             return $this->redirectToRoute('admin_album_index');
         }
 
-        return $this->render('admin/album/add.html.twig', ['form' => $form->createView()]);
+        return $this->render('admin/album/add.html.twig', [
+            'form' => $form->createView(),
+        ]);
     }
 
     #[Route('/update/{id}', name: 'admin_album_update')]
-    public function update(Request $request, int $id)
+    public function update(Request $request, int $id): Response
     {
         $album = $this->albumRepository->find($id);
+
+        if (!$album) {
+            $this->addFlash('error', 'Album introuvable.');
+            return $this->redirectToRoute('admin_album_index');
+        }
+
         $form = $this->createForm(AlbumType::class, $album);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->em->flush();
+            $this->addFlash('success', 'L’album a bien été mis à jour.');
 
             return $this->redirectToRoute('admin_album_index');
         }
 
-        return $this->render('admin/album/update.html.twig', ['form' => $form->createView()]);
+        return $this->render('admin/album/update.html.twig', [
+            'form' => $form->createView(),
+        ]);
     }
 
     #[Route('/delete/{id}', name: 'admin_album_delete')]
-    public function delete(int $id)
+    public function delete(int $id): Response
     {
-        $media = $this->albumRepository->find($id);
-        $this->em->remove($media);
+        $album = $this->albumRepository->find($id);
+
+        if (!$album) {
+            $this->addFlash('error', 'Album introuvable.');
+            return $this->redirectToRoute('admin_album_index');
+        }
+
+        $medias = $this->mediaRepository->findBy(['album' => $album]);
+
+        foreach ($medias as $media) {
+            $filePath = $this->getParameter('kernel.project_dir') . '/public/' . $media->getPath();
+
+            if (file_exists($filePath) && is_file($filePath)) {
+                unlink($filePath);
+            }
+
+            $this->em->remove($media);
+        }
+
+        $this->em->remove($album);
         $this->em->flush();
+
+        $this->addFlash('success', 'L’album et ses médias associés ont bien été supprimés.');
 
         return $this->redirectToRoute('admin_album_index');
     }
